add new flag to use openblas for SVD
It was discovered that the default SVD provided by `nalgebra` has a
known "occasional loss of accuracy" problem in some cases (see link
below), which translates to an element-wise error of around `1e-4` when
the decomposition is recombined and compared to the original matrix.
This causes problems for this project because it introduces a `1e-4`
noise floor in the main quantity we want to look at in `workspace`

This commit introduces an optional dependency on `nalgebra-lapack` to
use OpenBLAS for SVDs instead (under feature flag `openblas`).
Unfortunately, `nalgebra-lapack` isn't as sophisticated as, say,
`ndarray-linalg`, and provides *only* a real-valued SVD with `f64`s. To
get around this, the `openblas` flag adds a few `f64` conversion methods
to `ComplexFloatExt` in order to perform SVDs using the well-known trick
of encoding the data of a complex-valued matrix `A` into a larger,
purely real-valued matrix `A'` with
```
A' = [ Re(A), -Im(A) ; Im(A), Re(A) ]
```
(see `mps::Schmidt::from_decomp` for more details). However, this
strategy is somewhat more error-prone and introduces additional overhead
from processing a larger matrix and having to encode/decode complex
values. From some minimal testing, this introduces a ~20x reduction in
speed, so one may as well switch back to the `ndarray` implementations
at that point.

https://github.com/dimforge/nalgebra/issues/1172